requirements/production.txt
Flask
flask-sqlalchemy
psycopg2-binary
flask-migrate
wheel
python-ucam-webauth
